api/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# AWS
boto3==1.34.34
botocore==1.34.34

# Validation and utilities
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1
python-dotenv==1.0.1

# Image processing (for thumbnails)
Pillow==10.2.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0

# Development
black==24.1.1
ruff==0.2.0
